Rename Rust file within same name directory to fix "module-inception"
